backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
python-multipart
tavily-python
openai
python-dotenv
